Library /Library_03032023_2

+++ src/main/java/com/skypro/library/service/BookServiceImpl.java
@@ -3,8 +3,9 @@ package com.skypro.library.service;
 import com.skypro.library.dao.BookDAO;
 import com.skypro.library.entity.Books;
 import org.springframework.stereotype.Service;
+import com.skypro.library.exceptions.*;
+import org.springframework.transaction.annotation.Transactional;
 
-import java.awt.print.Book;
 import java.util.List;
 
 @Service
@@ -19,39 +20,43 @@ public class BookServiceImpl implements BookService {
         this.bookDAO = bookDAO;
     }
 
-    @Override                   //1. Создание книги с помощью запроса INSERT
-    public void addEBookService(Books book) {
-        book.setIsbn(null);
-        bookDAO.addEBookDAO(book);
+    @Override
+    @Transactional                                  //1. Создание книги с помощью запроса INSERT
+    public void addBookService(Books book) {
+        bookDAO.addBookDAO(book);
 
     }
 
-    @Override               //2. Изменение названия, автора или года издания книги с помощью одного запроса UPDATE который принимает все три параметра
+    @Override
+    @Transactional                                              //2. Изменение названия, автора или года издания книги с помощью одного запроса UPDATE который принимает все три параметра
     public void updateBookService(Books book) {
         bookDAO.updateBookDAO(book);
     }
 
-    @Override                           //3. Удаление книги по ISBN
-    public void deleteBookService(Character isbn) {
+    @Override
+    @Transactional                                                    //3. Удаление книги по ISBN
+    public void deleteBookService(String isbn) {
         Books book = bookDAO.getBookByIsbnDAO(isbn);
-//        if (book == null) {
-//            throw new BookException("Person with id = " + id + " doesn't exist");
-//        }
+        if (book == null) {
+            throw new BookException("Person with id = " + isbn + " doesn't exist");
+        }
 
         bookDAO.deleteBookDAO(isbn);
     }
 
-    @Override                       //4. Получение всех книг
+    @Override
+    @Transactional                                //4. Получение всех книг
     public List<Books> getBooksService() {
         return bookDAO.getBooksDAO();
     }
 
-    @Override                          //5. Получение одной книги по ISBN
-    public Books getBookByIsbnService(Character isbn) {
+    @Override
+    @Transactional                              //5. Получение одной книги по ISBN
+    public Books getBookByIsbnService(String isbn) {
         Books book = bookDAO.getBookByIsbnDAO(isbn);
-//        if (book == null) {
-//            throw new BookException("Person with id = " + id + " doesn't exist");
-//        }
+        if (book == null) {
+            throw new BookException("Person with isbn = " + isbn + " doesn't exist");
+        }
         return book;
     }
 }
